Create 1. Set Matrix Zeroes
https://leetcode.com/problems/set-matrix-zeroes/

This problem asks us to set entire rows and columns to zero if any cell within them is zero. The intuition is that we can scan the matrix to find which rows and columns need to be zeroed, using the first row and first column as markers (rather than extra space). By storing “zero flags” in the first row and column for any cell that is zero, we can then perform a second pass over the matrix and set the appropriate rows and columns to zero. Finally, we zero out the first row and first column themselves if they were flagged, ensuring we only use constant additional space.