Create 8. Merge Intervals
https://leetcode.com/problems/merge-intervals/

This solution merges overlapping intervals by first sorting them according to their start times, then iterating through the sorted list and comparing each interval to the last merged interval. If an interval’s start time is within or immediately follows the last merged interval’s end time, the two intervals overlap; in that case, we update the end time of the last merged interval to be the maximum of both intervals’ end times. Otherwise, we simply add the new interval to the list of merged intervals. This process ensures all overlapping intervals are merged efficiently in a single pass after sorting.