Create 21. Middle of the Linked List
This solution efficiently finds the middle node of a singly linked list using the **two-pointer technique**. It uses a `fast` pointer that moves two steps at a time and a `slow` pointer that moves one step at a time. By the time the `fast` pointer reaches the end of the list, the `slow` pointer will be at the middle node. This approach ensures \(O(n)\) time complexity since each pointer traverses the list at most once, and \(O(1)\) space complexity as no additional data structures are used. If there are two middle nodes (in the case of an even-length list), the second middle node is returned.