Create 25. Delete Node in a Linked List
This solution efficiently deletes a given node (except the tail) from a singly linked list by copying the value of the next node into the current node and then bypassing the next node in the list. Instead of actually removing the node itself, this approach modifies the current node to "become" the next node, effectively deleting it from the perspective of the list. This method works in \(O(1)\) time and requires no additional space, making it very efficient. However, it relies on the fact that the node to be deleted is not the tail node.